moved componentDidMount functions for todos to todos component
